1. 将boll参数改为浮点数。
git-svn-id: file:///d/clouddrive/SyncFolder/svnRepository/trunk@139 3ddc6574-91de-f842-91ed-6e7e04a71137

+++ TradeAgentVS/TradeAgentVS/src/Stratergy/BollingerBand.h
@@ -43,7 +43,7 @@ public:
 	{
 		return (mIndex>mBollPeriod+10)&&mDataValid;
 	}
-	bool CalcBoll(double aLastPrice, int aPeriod, int aOutterAmp, int aInnerAmp)
+	bool CalcBoll(double aLastPrice, int aPeriod, double aOutterAmp, double aInnerAmp)
 	{
 		BollingerBandData lTempBollData;
 		mBollPeriod = aPeriod;
